+++ certbot/CHANGELOG.md
@@ -13,6 +13,10 @@ Certbot adheres to [Semantic Versioning](https://semver.org/).
 * `acme.client.ClientBase`, `acme.messages.Authorization.resolved_combinations`
   `acme.messages.Authorization.combinations` and `acme.mixins` are deprecated and
   will be removed in a future release.
+* The `certbot-dns-cloudxns` plugin is now deprecated and will be removed in the
+  next major release of Certbot.
+* The `source_address` argument for `acme.client.ClientNetwork` is deprecated
+  and support for it will be removed in the next major release.
 
 ### Fixed